reafactor: 重构脚本提高代码复用率
