+++ src/BootstrapBlazor.Server/wwwroot/lib/theme.js
@@ -1,4 +1,3 @@
-﻿import { getPreferredTheme } from "../../_content/BootstrapBlazor/modules/theme.js"
-import { setTheme } from "../../_content/BootstrapBlazor/modules/utility.js"
+﻿import { getPreferredTheme, setTheme } from "../../_content/BootstrapBlazor/modules/utility.js"
 
-setTheme(getPreferredTheme())
+setTheme(getPreferredTheme(), false)
